update test package

+++ packages/lucia-sveltekit/src/auth/endpoints/logout.ts
@@ -2,18 +2,17 @@ import type { RequestEvent } from "../../kit.js";
 import { createBlankCookies } from "../../utils/cookie.js";
 import type { Context } from "../index.js";
 import { ErrorResponse } from "./index.js";
-import type { LuciaError } from "../../utils/error.js";
+import { LuciaError } from "../../utils/error.js";
 
 export const handleLogoutRequest = async (
     event: RequestEvent,
     context: Context
 ) => {
     try {
-        const session = await context.auth.validateRequest(event.request);
-        const [accessToken] = session.accessToken;
-        const [refreshToken] = session.refreshToken;
+        const { accessToken, refreshToken } = await context.auth.parseRequest(event.request)
+        if (!accessToken) throw new LuciaError("AUTH_INVALID_ACCESS_TOKEN")
         await Promise.allSettled([
-            context.adapter.deleteRefreshToken(refreshToken),
+            context.adapter.deleteRefreshToken(refreshToken || ""),
             context.adapter.deleteSessionByAccessToken(accessToken),
         ]);
         return new Response(null, {
